% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cherry_reconstruction.R
\name{add_pseudonode}
\alias{add_pseudonode}
\title{Add pseudo node}
\usage{
add_pseudonode(phylo)
}
\arguments{
\item{phylo}{an object of class "phylo".}
}
\value{
a phylo structure
}
\description{
A function which adds a node to the root which allows flexible observation time
}
