% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_reconstruction.R
\name{bifur_punish}
\alias{bifur_punish}
\title{Topology biased score}
\usage{
bifur_punish(bifurcation_pro, edgelength)
}
\arguments{
\item{bifurcation_pro}{a parameter which describes the proportion of cell not bifurcated after one generation time.}

\item{edgelength}{a data frame which bind the edge attribute and edge.length attribute of a phylo structure.}
}
\value{
a length-one numeric
}
\description{
This function takes the proportion of  bifurcation event/non bifurcation event
as the base and number of non-bifurcation event as the power to propose a topology biased
score under the discrete edge length and fixed tree height model in which
edge length bigger than one implies (n-1) continuous non bifurcation event.
}
