% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cherry_reconstruction.R
\name{cherry_based_recon}
\alias{cherry_based_recon}
\title{Tree Reconstruction}
\usage{
cherry_based_recon(
  character_info,
  mu,
  alpha,
  non_bifur_pro,
  alternative_threshold,
  site_num,
  state_num,
  precluster_replicate = TRUE
)
}
\arguments{
\item{character_info}{A data frame with two columns which saves cell name and barcode respectively}

\item{mu}{a vector of site specific mutation probability.}

\item{alpha}{a list of vectors which describe the site specific priors of mutation outcomes.}

\item{non_bifur_pro}{A parameter which describes the proportion of cell not bifurcated after one generation time}

\item{alternative_threshold}{Default as zero, if 1, two sequences are 10 times more likely to have a bifurcation event in 2 generation.}

\item{site_num}{the length of barcode}

\item{state_num}{number of mutation outcomes}

\item{precluster_replicate}{Logical. whether or not to cluster the cells with identical barcodes as cluster before reconstruct phylogeny.default as TRUE.}
}
\value{
a phylo structure without edge length
}
\description{
This function reconstruct the tree based on the pairwise likelihood of having one bifurcation in two generation time under the bottom up approach.
The barcode of unobserved internal vertex follows the irreversible mutation rule and the fixed root state.
}
