% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edgelength_assignment.R
\name{cherry_likelihood}
\alias{cherry_likelihood}
\title{Likelihood on cherry structure}
\usage{
cherry_likelihood(
  left_child,
  right_child,
  mu,
  alpha,
  non_bifur_pro,
  branch_l,
  branch_r,
  r_height
)
}
\arguments{
\item{left_child}{n(length of site) x m(number of states in one site) binary matrix which saves the barcode of left node of a cherry}

\item{right_child}{n(length of site) x m(number of states in one site) binary matrix which saves the barcode of right node of a cherry}

\item{mu}{a vector of site specific mutation probability.}

\item{alpha}{a list of vectors which describe the site specific priors of mutation outcomes.}

\item{non_bifur_pro}{a parameter which describe the proportion of cell not bifurcated after one generation time.}

\item{branch_l}{branch length of left child}

\item{branch_r}{branch length of right child}

\item{r_height}{branch length incoming edges connecting the common ancestor node of two cherry to the fixed root}
}
\value{
a likelihood score on the cherry structure
}
\description{
Likelihood on cherry structure
}
