% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_metric.R
\name{colless_index}
\alias{colless_index}
\title{Compute Colless imbalance}
\usage{
colless_index(tree)
}
\arguments{
\item{tree}{a phylogenetic tree}
}
\value{
colless imbalance value
}
\description{
Compute Colless imbalance
}
