% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_metric.R
\name{colless_index_weighted}
\alias{colless_index_weighted}
\title{Weighted colless index}
\usage{
colless_index_weighted(tree)
}
\arguments{
\item{tree}{a phylogenetic tree}
}
\value{
colless imbalance value with the weight descirbed in the reference paper
}
\description{
Weighted colless index
}
