% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nni_deltree.R
\name{deltree_likelihood}
\alias{deltree_likelihood}
\title{Likelihood computation}
\usage{
deltree_likelihood(
  discrete_EdgeLength_tree,
  state_num,
  non_bifur_pro,
  mu,
  alpha
)
}
\arguments{
\item{discrete_EdgeLength_tree}{an object of class "phylo"}

\item{state_num}{number of mutation outcomes}

\item{non_bifur_pro}{A parameter which describes the proportion of cell not bifurcated after one generation time}

\item{mu}{a vector of site specific mutation probability.}

\item{alpha}{a list of vectors which describe the site specific priors of mutation outcomes.}
}
\value{
a vector which stores the node within a subtree full of replicate barcode sequence
}
\description{
Likelihood computation
}
