% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nni_deltree.R
\name{edgelength_inter}
\alias{edgelength_inter}
\title{Bottom up edge length iteration which reduce non bifurcation event}
\usage{
edgelength_inter(edgelength, free_node_initial)
}
\arguments{
\item{edgelength}{a data frame which bind the edge attribute and edge.length attribute of a phylo structure.}

\item{free_node_initial}{index of parent node whose two descendant node both have non bifurcation events}
}
\value{
a data frame which bind the edge attribute and edge.length attribute of a phylo structure.
}
\description{
Bottom up edge length iteration which reduce non bifurcation event
}
