% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nni_move_with_limited_ndoe.R
\name{emurate_node_path_withpseudonode}
\alias{emurate_node_path_withpseudonode}
\title{Likelihood score}
\usage{
emurate_node_path_withpseudonode(
  phylo,
  parent.node,
  node_edge_length,
  mu,
  alpha
)
}
\arguments{
\item{phylo}{an object of class "phylo".}

\item{parent.node}{a list of n(length of site) x m(number of states in one site) binary matrix which saves the barcode of leaves}

\item{node_edge_length}{a data frame which bind the edge attribute and edge.length attribute of a phylo structure.}

\item{mu}{a vector of site specific mutation probability.}

\item{alpha}{a list of vectors which describe the site specific priors of mutation outcomes.}
}
\value{
A list of matrix of number of nodes
}
\description{
Likelihood score
}
