% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_reconstruction.R
\name{get_consensus}
\alias{get_consensus}
\title{Title}
\usage{
get_consensus(node_info, site_index, site_num)
}
\arguments{
\item{node_info}{A n(cell_num) x m(site_num) matrix which stores the barcode}

\item{site_index}{The index of two descendant cell of a node}

\item{site_num}{The length of barcode}
}
\value{
A vector of with the length of site_num which stores the barcode of the parent node
}
\description{
This function applies a parsimony method to assign the barcode with the irreversible rule of mutation and the barcode of its two direct descendant.
}
