% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nni_move_with_limited_node.R
\name{get_nni_node_noRT}
\alias{get_nni_node_noRT}
\title{Selet nni node}
\usage{
get_nni_node_noRT(phylo)
}
\arguments{
\item{phylo}{an object of class "phylo"}
}
\value{
a list of nni trees
}
\description{
Selet nni node
}
