% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edgelength_assignment.R
\name{get_node_character}
\alias{get_node_character}
\title{Assign barcode of unobserved node with parsimony}
\usage{
get_node_character(phylo_c)
}
\arguments{
\item{phylo_c}{an object of class "phylo"}
}
\value{
a dataframe which stores the barcode of all nodes including observed leaves and unobserved internal node
}
\description{
Assign barcode of unobserved node with parsimony
}
