% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_reconstruction.R
\name{get_parent}
\alias{get_parent}
\title{Likelihood computation with matrix}
\usage{
get_parent(
  child_left,
  child_right,
  left_branch_length,
  right_branch_length,
  mu,
  alpha
)
}
\arguments{
\item{child_left}{n(length of site) x m(number of states in one site) binary matrix which saves the barcode of left node of a cherry}

\item{child_right}{n(length of site) x m(number of states in one site) binary matrix which saves the barcode of right node of a cherry}

\item{left_branch_length}{left branch length}

\item{right_branch_length}{right branch length}

\item{mu}{a vector of site specific mutation probability.}

\item{alpha}{a list of vectors which describe the site specific priors of mutation outcomes.}
}
\value{
n(length of site) x m(number of states in one site) binary matrix which saves the barcode of common ancestor node of a cherry
}
\description{
Likelihood computation with matrix
}
