% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_reconstruction.R, R/simulation.R
\name{get_probmatrix}
\alias{get_probmatrix}
\title{Probability matrix of dicrete edge length}
\usage{
get_probmatrix(t, mu, alpha, state_num)

get_probmatrix(t, mu, alpha, state_num)
}
\arguments{
\item{t}{integer.number of generation or number of bifurcation}

\item{mu}{a vector of site specific mutation probability.}

\item{alpha}{a list of vectors which describe the site specific priors of mutation outcomes.}

\item{state_num}{number of mutation outcomes.}
}
\value{
A probability matrix

A probability matrix
}
\description{
Probability matrix of dicrete edge length

Probability matrix of dicrete edge length
}
