% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{get_site_mutation}
\alias{get_site_mutation}
\title{Count the number of mutational events}
\usage{
get_site_mutation(phylo_c, site_num, state_num)
}
\arguments{
\item{phylo_c}{an object of class "phylo"}

\item{site_num}{the length of barcode}

\item{state_num}{number of possible mutation outcomes}
}
\value{
a matrix which stores the number of mutational/non mutation events on edge of the tree
}
\description{
Count the number of mutational events
}
