% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nni_deltree.R
\name{get_sparenode_name}
\alias{get_sparenode_name}
\title{Find non-bifurcation node}
\usage{
get_sparenode_name(edgelength)
}
\arguments{
\item{edgelength}{a data frame which bind the edge attribute and edge.length attribute of a phylo structure.}
}
\value{
a vector of node index
}
\description{
Find non-bifurcation node
}
