% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nni_deltree.R
\name{initial_edgelength_pseudonode}
\alias{initial_edgelength_pseudonode}
\title{Initial edgelength assignment with longest pending edge}
\usage{
initial_edgelength_pseudonode(phylo, nGen)
}
\arguments{
\item{phylo}{an object of class "phylo".}

\item{nGen}{fixed tree height based on experimental duration and one generation of certain cell}
}
\value{
a data frame which bind the edge attribute and edge.length attribute of a phylo structure.
}
\description{
Initial edgelength assignment with longest pending edge
}
