% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{lineage_sim}
\alias{lineage_sim}
\title{Lineage simulation}
\usage{
lineage_sim(tree, state_num, site_num, mu, alpha)
}
\arguments{
\item{tree}{an object of class "phylo".}

\item{state_num}{number of possible mutation outcomes}

\item{site_num}{the length of barcode}

\item{mu}{a vector of site specific mutation probability.}

\item{alpha}{a list of vectors which describe the site specific priors of mutation outcomes.}
}
\value{
a data frame with two columns which saves cell name and barcode of leaves respectively
}
\description{
Lineage simulation
}
