% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nni_move_with_limited_node.R
\name{local_optiaml_withpseudoRT}
\alias{local_optiaml_withpseudoRT}
\title{Bottom up edgelength optimzation}
\usage{
local_optiaml_withpseudoRT(tree, mu, alpha, nGen, non_bifur_pro, state_num)
}
\arguments{
\item{tree}{an object of class "phylo"}

\item{mu}{a vector of site specific mutation probability.}

\item{alpha}{a list of vectors which describe the site specific priors of mutation outcomes.}

\item{nGen}{fixed tree height based on experimental duration and one generation of certain cell}

\item{non_bifur_pro}{A parameter which describes the proportion of cell not bifurcated after one generation time}

\item{state_num}{number of mutation outcomes}
}
\value{
a data frame which bind the edge attribute and edge.length attribute of a phylo structure.
}
\description{
Bottom up edgelength optimzation
}
