% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nni_move_with_limited_node.R
\name{nj_tree}
\alias{nj_tree}
\title{Title Neighbor-joining with define hamming distance,especially for barcode with three state and "1" denoted as original state}
\usage{
nj_tree(character_info, site_num)
}
\arguments{
\item{character_info}{A data frame with two columns which saves cell name and barcode respectively}

\item{site_num}{the length of barcode}
}
\value{
a rooted binary tree.
}
\description{
Title Neighbor-joining with define hamming distance,especially for barcode with three state and "1" denoted as original state
}
