% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nni_deltree.R
\name{nni_tree_noRT}
\alias{nni_tree_noRT}
\title{Tree rearrangements of one NNI(nearest neighbor interchange) move without node within certain cluster}
\usage{
nni_tree_noRT(phylo)
}
\arguments{
\item{phylo}{an object of class "phylo".}
}
\value{
a list of nni trees
}
\description{
Tree rearrangements of one NNI(nearest neighbor interchange) move without node within certain cluster
}
