% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{node_state_sim}
\alias{node_state_sim}
\title{Lineage barcode simulation}
\usage{
node_state_sim(branch_length, parent_state, mu, alpha, site_num, state_num)
}
\arguments{
\item{branch_length}{branch length}

\item{parent_state}{n(length of site) x m(number of states in one site) binary matrix which saves the barcode of parent node}

\item{mu}{a vector of site specific mutation probability.}

\item{alpha}{a list of vectors which describe the site specific priors of mutation outcomes.}

\item{site_num}{the length of barcode}

\item{state_num}{number of possible mutation outcomes}
}
\value{
a binary matrix which stores the barcode
}
\description{
Lineage barcode simulation
}
