% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_reconstruction.R
\name{onehot_coding}
\alias{onehot_coding}
\title{Title}
\usage{
onehot_coding(processed_tip_label, state_num)
}
\arguments{
\item{processed_tip_label}{barcode sequence with a prefix}

\item{state_num}{number of mutation outcomes}
}
\value{
A list with the length of cell number. each element include a site_num x state_num binary matrix
}
\description{
By default, this function take a matrix which stores the barcode information of observed cell and convert the barcode sequence into m(length of the barcode or mutation site) x n(number of mutation outcomes) binary matrix with one hot encoding techniques,
the binary matrix of barcode are saved in the list of n cells.
}
