% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_reconstruction.R
\name{prefix_state}
\alias{prefix_state}
\title{Title}
\usage{
prefix_state(node_info, state_num)
}
\arguments{
\item{node_info}{A n(cell_num) x m(site_num) matrix which stores the barcode}

\item{state_num}{number of mutation outcomes}
}
\value{
the barcode sequence with a prefix which contains all possible mutation outcomes
}
\description{
A preprocess step which adds a prefix which contains all possible mutation outcomes to ensure the dimention of binary matrix before one hot encoding
}
