% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edgelength_assignment.R
\name{remove_pseudonode}
\alias{remove_pseudonode}
\title{Remove the pseudonode added}
\usage{
remove_pseudonode(phylo)
}
\arguments{
\item{phylo}{an object of "phylo"}
}
\value{
phylo
}
\description{
Remove the pseudonode added
}
