% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{reverse_onehot_encoding}
\alias{reverse_onehot_encoding}
\title{reverse onehot encoding}
\usage{
reverse_onehot_encoding(n_state_matrix, state_num)
}
\arguments{
\item{n_state_matrix}{n(length of site) x m(number of states in one site) binary matrix which saves the barcode}

\item{state_num}{number of possible mutation outcomes}
}
\value{
a vector of barcode
}
\description{
reverse onehot encoding
}
