% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_reconstruction.R
\name{str_cherry_lik}
\alias{str_cherry_lik}
\title{Likelihood of bifurcation within two generations}
\usage{
str_cherry_lik(left_child, right_child, mu, alpha, non_bifur_pro)
}
\arguments{
\item{left_child}{n(length of site) x m(number of states in one site) binary matrix which saves the barcode of left node of a cherry}

\item{right_child}{n(length of site) x m(number of states in one site) binary matrix which saves the barcode of right node of a cherry}

\item{mu}{a vector of site specific mutation probability.}

\item{alpha}{a list of vectors which describe the site specific priors of mutation outcomes.}

\item{non_bifur_pro}{a parameter which describe the proportion of cell did not bifurcation after one generation time.}
}
\value{
a non-negative score which describe  how many times its more likely to bifurcate in one generation time than not to bifurcate.
}
\description{
this function compute a pairwise likelihood of two cells orgins from a common ancestor within one bifurcation time. Given a tree height 2,
there is an optional unobserved ancestor assiociated with an bifurcation event in the cell. The subtraction of having and
not having an internal node are used as the pairwise distance of two barocdes in two generation time.
}
