% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{topology_simulation}
\alias{topology_simulation}
\title{a forward approach to simulate tree topology with bifurcation}
\usage{
topology_simulation(bifur_pro, nGen)
}
\arguments{
\item{bifur_pro}{a paramter which describe the proportion of cell bifurcated after one generation time.}

\item{nGen}{fixed tree height of n generation time}
}
\value{
an object of class "phylo"
}
\description{
a forward approach to simulate tree topology with bifurcation
}
