% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nni_move_with_limited_node.R
\name{tree_likelihood}
\alias{tree_likelihood}
\title{Likelihood computation}
\usage{
tree_likelihood(discrete_EdgeLength_tree, state_num)
}
\arguments{
\item{discrete_EdgeLength_tree}{an object of class "phylo"}

\item{state_num}{number of mutation outcomes}
}
\value{
a vector which stores the node within a subtree full of replicate barcode sequence
}
\description{
Likelihood computation
}
