% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_metric.R
\name{triplet_distance_weighted}
\alias{triplet_distance_weighted}
\title{Weighted Tree Metric}
\usage{
triplet_distance_weighted(tree1, tree2, normalized = TRUE)
}
\arguments{
\item{tree1}{An oject of class "phylo"}

\item{tree2}{An oject of class "phylo"}

\item{normalized}{default set to TRUE; to normalize the triplet distance by the total number of possible triplets.}
}
\value{
triplet distance between the two trees
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This function extends the triplet_distance function by incorporating information about the branch lengths of the phylogenetic trees.
For matching triplets(identical topology),the weight is calculated as the absolute difference in edge lengths between the common ancestor of the triplet in tree1 and tree2.
For Non-matching Triplets,the weight is calculated as the average edge length of the common ancestor in both trees.
}\if{html}{\out{</div>}}
}
