% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_reconstruction.R
\name{unique_state}
\alias{unique_state}
\title{Precluster of replicates}
\usage{
unique_state(charatcer_info)
}
\arguments{
\item{charatcer_info}{an dataframe with two columns which saves cell name and barcode respectively}
}
\value{
a list with three attribute:cluster_topo,barcode,round
}
\description{
This function remove the stocastic influence of replicates by clustering the replicates  in advance.
An atrribute 'round' was set as the minimum tree height of the cluster for sake of the fixed tree.
for example, a cluster with one cell has a round atrribute of 1, and a cluster of 6 cell has a round attribute of 4.
After the preprocess step, a list with three attribute:cluster_topo,barcode,round will be new candicates for tree reconstruction.
}
