% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nni_move_with_limited_ndoe.R
\name{within_replicate_node}
\alias{within_replicate_node}
\title{Identify node within the replicate cluster}
\usage{
within_replicate_node(phylo)
}
\arguments{
\item{phylo}{an object of class "phylo"}
}
\value{
a vector which stores the node within a subtree full of replicate barcode sequence
}
\description{
Identify node within the replicate cluster
}
