% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edgelength_assignment.R
\name{direct_assignment}
\alias{direct_assignment}
\title{a bottom-up approach to assign the edge length under fixed tree height}
\usage{
direct_assignment(phylo, non_bifur_pro, nGen, state_num)
}
\arguments{
\item{phylo}{an object of class "phylo"}

\item{non_bifur_pro}{a parameter which describe the proportion of cell not bifurcated after one generation time.}

\item{nGen}{fixed tree height based on experimental duration and one generation of certain cell}

\item{state_num}{number of possible mutation outcomes}
}
\value{
an object of class "phylo"
}
\description{
a bottom-up approach to assign the edge length of a phylo object.The edge length will be the
subtraction of vertex height of two node. At first, the vertex height of leaves are set
as the fixed tree height, then the cherry structure is assigned an optimal
tree height as the vertex height. We assume an peak in likelihood with fixed tree height.
}
