% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edgelength_assignment.R,
%   R/parameter_estimation.R
\name{get_node_character}
\alias{get_node_character}
\title{Assign barcode of unobserved node with parsimony}
\usage{
get_node_character(phylo_c)

get_node_character(phylo_c)
}
\arguments{
\item{phylo_c}{a object of "phylo"}
}
\value{
a dataframe which stores the barcode of all nodes including observed leaves and unobserved internal node

a dataframe which stores the the barcode of leaves and nodes
}
\description{
Assign barcode of unobserved node with parsimony

Assign barcode with Parsimony
}
