% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nni_move_with_limited_ndoe.R
\name{nni_iter_withedgelength_pseudonode}
\alias{nni_iter_withedgelength_pseudonode}
\title{NNI move and edge length optimization}
\usage{
nni_iter_withedgelength_pseudonode(
  current_tree,
  mu,
  alpha,
  nGen,
  non_bifur_pro,
  state_num
)
}
\arguments{
\item{current_tree}{an object of class "phylo".}

\item{mu}{a vector of site specific mutation probability.}

\item{alpha}{a list of vectors which describe the site specific priors of mutation outcomes.}

\item{nGen}{fixed tree height based on experimental duration and one generation of certain cell}

\item{non_bifur_pro}{A parameter which describes the proportion of cell not bifurcated after one generation time}

\item{state_num}{number of mutation outcomes}
}
\value{
a list of nni trees with optimized edge length and likelihood.
}
\description{
NNI move and edge length optimization
}
